refactor: Adapt list_parts return type to u64

+++ components/data_proxy/src/s3_frontend/s3service.rs
@@ -584,7 +584,7 @@ impl S3 for ArunaS3Service {
             is_truncated,
             key: Some(req.input.key),
             max_parts: req.input.max_parts,
-            next_part_number_marker: next_part_number_marker.map(|marker| marker as i32),
+            next_part_number_marker: next_part_number_marker.map(|marker| marker),
             owner: None, //Todo?
             part_number_marker: req.input.part_number_marker,
             parts,
